Update: HistoryActivity Refactor + OngoingActivity Refactor
