+++ Serabutan/Features/View/Activity/Ongoing/OngoingActivityVC.swift
@@ -7,7 +7,7 @@
 
 import UIKit
 
-class OngoingActivityVC: UIViewController, UITableViewDelegate, UITableViewDataSource {
+class OngoingActivityVC: UIViewController {
     
     @IBOutlet weak var ongoingActivityTable: UITableView!
     
@@ -36,43 +36,4 @@ class OngoingActivityVC: UIViewController, UITableViewDelegate, UITableViewDataS
             self.ongoingActivityTable.reloadData()
         }
     }
-}
-
-extension OngoingActivityVC {
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return dummyData.count
-    }
-    
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = ongoingActivityTable.dequeueReusableCell(withIdentifier: AssistanceTableViewCell.identifier, for: indexPath) as! AssistanceTableViewCell
-        cell.selectionStyle = UITableViewCell.SelectionStyle.none
-        let data = dummyData[indexPath.row]
-        let poster = data.jobPosterId
-        
-        cell.tagView.isHidden = true
-        cell.availableView.isHidden = false
-        
-        let newView = cell.availabelInsetView1
-        newView!.heightAnchor.constraint(equalToConstant: 50).isActive = true
-        
-        cell.helperView.isHidden = true
-        cell.youHelperView.isHidden = true
-        
-        cell.setStatusView(urgency: data.urgency)
-        cell.titleLabel.text = data.title
-        cell.headerLabel.text = StringFormatter().distance(data.distance)
-        cell.compensationLabel.text = StringFormatter().priceFormatting(amount: data.price)
-        cell.posterImage.image = poster.avatar
-        cell.posterLabel.text = poster.name
-        cell.timeElapsedLabel.text = StringFormatter().relativeDateFormatter(date: data.postingDate)
-        cell.verifiedLogo.isHidden = !data.jobPosterId.isVerified
-        return cell
-    }
-    
-    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        let detailBantuan = DetailBantuanVC()
-        detailBantuan.selectedJob = dummyData[indexPath.row]
-        self.navigationController?.pushViewController(detailBantuan, animated: true)
-    }
-    
 }
